chore: rename data-node node to data-node start + remove ulimits
