+++ cmd/data-node/start/node_post.go
@@ -10,7 +10,7 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package node
+package start
 
 import (
 	"strings"
